requirements.txt
python_version == 3.10.11
pillow == 10.1.0
django == 4.2.6
django-rest-framework == 3.14.0
mysqlclient == 2.2.0
django-taggit == 4.0.0
djangorestframework-simplejwt == 5.3.0
django-ckeditor == 6.7.0
django-rest-passwordreset == 1.3.0
beautifulsoup4 == 4.12.2
drf-yasg == 1.21.7